+++ test/cram/show/admin_default.t
@@ -1,4 +1,7 @@
 
-  $ caretaker show --admin-dir ../admin --source admin
+  $ caretaker show \
+  >   --source admin \
+  >   --admin-dir ../admin \
+  >   --data-dir ../data
   caretaker: [ERROR] invalid source: cannot read project from admin
   [1]