+++ test/cram/show/admin_heatmap.t
@@ -1,5 +1,9 @@
 
-  $ caretaker show --admin-dir ../admin --source admin --heatmap
+  $ caretaker show \
+  >   --source admin \
+  >   --admin-dir ../admin \
+  >   --data-dir ../data \
+  >   --heatmap
   KR123        [19.875] 2022-10-02 => 2022-10-30
   -----------
   Total        [19.875]