+++ test/cram/show/local_default.t
@@ -1,5 +1,36 @@
-  $ caretaker show --source local --data-dir ../data
+
+  $ caretaker show \
+  >   --source local \
+  >   --data-dir ../data
   org: tarides
   
   == Project (25) ==
+  
+  $ caretaker show \
+  >   --source local \
+  >   --data-dir ../data \
+  >   --number 27
+  org: tarides
+  
+  == <title> (27) ==
+    [  KR123] <title> (open)
+      Status    : <status>
+      Quarter   : <quarter>
+      Starts    : <starts>
+      Ends      : <ends>
+      Team      : <team>
+      Funder    : <funder>
+      effort days*: 117.
+      junior fte*: 2.6
+      senior fte*: 10.4
+      principal fte*: 7.8
+      end on quarter*: <end-on-quarter>
+      duration (weeks)*: 24.
+      start on quarter*: <start-on-quarter>
+      js bucket*: <js-bucket>
+      contact*  : <contact>
+      owner *   : <owner>
+      proposal link*: <proposal-link>
+      proposal status*: Active
+      repository*: <repository>