+++ test/cram/show/local_heatmap.t
@@ -1,4 +1,8 @@
-  $ caretaker show --source local --data-dir ../data --heatmap
+
+  $ caretaker show \
+  >   --source local \
+  >   --data-dir ../data \
+  >   --heatmap
   KR123        [21] 2022-10-02 => 2022-10-30
   -----------
   Total        [21]