+++ test/cram/show/local_timesheets.t
@@ -1,8 +1,11 @@
 
-  $ caretaker show --source local --data-dir ../data --timesheets
-  "Number","Id","Year","Month","Week","Days","Hours","User","Full Name","Funder","Entity Funder","Work Item","Team","Category","Objective"
-  "","KR123","2022","10","40","5","","eng1","","","","","","",""
-  "","KR123","2022","10","41","1","","eng1","","","","","","",""
-  "","KR123","2022","10","41","5","","eng1","","","","","","",""
-  "","KR123","2022","10","43","5","","eng1","","","","","","",""
-  "","KR123","2022","10","43","5","","eng1","","","","","","",""
+  $ caretaker show \
+  >   --source local \
+  >   --data-dir ../data \
+  >   --timesheets
+  "Number","Id","Year","Month","Week","Days","Hours","User","Full Name","Funder","Entity Funder","Objective","Team","Category"
+  "","KR123","2022","10","40","5","","eng1","","","","","",""
+  "","KR123","2022","10","41","1","","eng1","","","","","",""
+  "","KR123","2022","10","41","5","","eng1","","","","","",""
+  "","KR123","2022","10","43","5","","eng1","","","","","",""
+  "","KR123","2022","10","43","5","","eng1","","","","","",""